fix profiler`s showing state when game resume
